ai-video-analysis/requirements.txt

flask
opencv-python-headless
transformers
torch
pillow
gunicorn
